+++ src/components/Modals/index.js
@@ -1,15 +1,17 @@
 import ButtonModal from "./ButtonModal";
 import ContentModal from "./ContentModal";
 import FailModalContainer from "./FailModalContainer";
+import InputModal from "./InputModal";
 import InputModalContainer from "./InputModalContainer";
 import NextStageModalContainer from "./NextStageModalContainer";
 import ResetModalContainer from "./ResetModalContainer";
 
 export {
-  InputModalContainer,
-  ContentModal,
   ButtonModal,
+  ContentModal,
   FailModalContainer,
+  InputModal,
+  InputModalContainer,
   NextStageModalContainer,
   ResetModalContainer,
 };
